app/api/urls.py
from django.urls import include, path
from rest_framework import routers, serializers, viewsets
from . import views

router = routers.DefaultRouter()

urlpatterns = [
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace = 'rest_framework'))
]
